+++ changelog.md
@@ -1,11 +1,17 @@
 # 版本 & 更新日志
 ## Beta
+### 0.1.2
+【重要】修复`setData`在关闭缓存时报错的bug  
+`setData`、`removeData`为异步函数了  
+更改`onTick`参数，改为`performanceImpact`（性能影响程度），移除`automaticTps`
+
 ### 0.1.1
 【重要】修复 `Storage Queue` bug  
 【重要】 修复 `gameLoop` bug  
 优化`Storage Queue`中`Map`的`key`  
 新增`Thing.use`方法  
 新增`Item.onUse`、`Item.onWear`、`Item.onDiswear`属性
+
 ### 0.1.0
 修复若干bug